+++ app/src/main/java/br/com/stone/emeraldcomponentsandroid/MainActivity.kt
@@ -2,7 +2,7 @@ package br.com.stone.emeraldcomponentsandroid
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import br.com.stone.emeraldcomponents.extension.setUp
 import br.com.stone.emeraldcomponentsandroid.activities.ButtonActivity
 import br.com.stone.emeraldcomponentsandroid.activities.CalendarActivity