+++ app/src/main/java/br/com/stone/emeraldcomponentsandroid/activities/DateActivity.kt
@@ -1,8 +1,8 @@
 package br.com.stone.emeraldcomponentsandroid.activities
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 import br.com.stone.emeraldcomponents.extension.format
 import br.com.stone.emeraldcomponentsandroid.R
 import kotlinx.android.synthetic.main.activity_date.*
@@ -27,5 +27,7 @@ class DateActivity : AppCompatActivity() {
                     startDate.format("dd/MM/yy") + " até " + endDate.format("dd/MM/yy")
                     , Toast.LENGTH_SHORT).show()
         }
+
+        emeraldDateFilter.maxDaysRange = 31
     }
 }