+++ app/src/main/java/br/com/stone/emeraldcomponentsandroid/activities/EditTextActivity.kt
@@ -1,9 +1,9 @@
 package br.com.stone.emeraldcomponentsandroid.activities
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 import br.com.stone.emeraldcomponentsandroid.R
 import kotlinx.android.synthetic.main.activity_emerald_edittext.*