+++ emeraldcomponents/src/main/java/br/com/stone/emeraldcomponents/basic/calendar/DateEventList.kt
@@ -3,14 +3,14 @@ package br.com.stone.emeraldcomponents.basic.calendar
 import android.content.Context
 import android.graphics.PorterDuff
 import android.graphics.PorterDuffColorFilter
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.DividerItemDecoration
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.LinearSmoothScroller
-import android.support.v7.widget.RecyclerView
 import android.util.AttributeSet
 import android.util.TypedValue
 import android.view.View
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.LinearSmoothScroller
+import androidx.recyclerview.widget.RecyclerView
 import br.com.stone.emeraldcomponents.R
 import br.com.stone.emeraldcomponents.basic.calendar.CalendarHelper.getGroupedEvents
 import br.com.stone.emeraldcomponents.basic.calendar.CalendarHelper.sortDates
@@ -100,7 +100,7 @@ class DateEventList : RecyclerView {
 
             emeraldEventList.setUp(dateEventList, { R.layout.event_list_content_item }, {
                 with(this.emeraldEventTitle) {
-                    text = it.spannable ?: ( it.title ?: "")
+                    text = it.spannable ?: (it.title ?: "")
                     setBackgroundColor(ContextCompat.getColor(context, it.color))
 
                     setTextColor(itemProperties.eventTitleTextColor)