+++ emeraldcomponents/src/test/java/br/com/stone/emeraldcomponents/basics/calendar/EmeraldCalendarViewTest.kt
@@ -1,11 +1,10 @@
 package br.com.stone.emeraldcomponents.basics.calendar
 
-import android.content.Context
+import androidx.test.core.app.ApplicationProvider
 import br.com.stone.emeraldcomponents.R
 import br.com.stone.emeraldcomponents.basic.calendar.EmeraldCalendarView
 import com.prolificinteractive.materialcalendarview.CalendarDay
 import org.junit.Assert
-import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.any
@@ -13,32 +12,25 @@ import org.mockito.Mockito.spy
 import org.mockito.Mockito.verify
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
-import org.robolectric.RuntimeEnvironment
 
 @RunWith(RobolectricTestRunner::class)
 class EmeraldCalendarViewTest {
-    private lateinit var context: Context
-
-    @Before
-    fun setup() {
-        context = RuntimeEnvironment.application
-    }
 
     @Test
     fun testInstanceWithContext() {
-        val view = EmeraldCalendarView(context)
+        val view = EmeraldCalendarView(ApplicationProvider.getApplicationContext())
         Assert.assertNotNull(view)
     }
 
     @Test
     fun testInstanceWithAttributeSet() {
-        val view = EmeraldCalendarView(context, Robolectric.buildAttributeSet().build())
+        val view = EmeraldCalendarView(ApplicationProvider.getApplicationContext(), Robolectric.buildAttributeSet().build())
         Assert.assertNotNull(view)
     }
 
     @Test
     fun testHighlightDays() {
-        val calendarView = spy(EmeraldCalendarView(context))
+        val calendarView = spy(EmeraldCalendarView(ApplicationProvider.getApplicationContext()))
         calendarView.highlightDays(R.color.emerald_textview_default_color, setOf(CalendarDay.today()))
         verify(calendarView).addDecorator(any())
     }