+++ emeraldcomponents/src/test/java/br/com/stone/emeraldcomponents/basics/calendar/EmeraldDateFilterTest.kt
@@ -1,21 +1,20 @@
 package br.com.stone.emeraldcomponents.basics.calendar
 
-import android.app.DatePickerDialog
-import android.support.v4.app.FragmentActivity
 import android.view.View
-import br.com.stone.emeraldcomponents.basic.calendar.EmeraldDateSelector
+import androidx.fragment.app.FragmentActivity
+import br.com.stone.emeraldcomponents.basic.calendar.EmeraldDateFilter
+import br.com.stone.emeraldcomponents.basic.calendar.EmeraldDateFilterOptions
 import br.com.stone.emeraldcomponents.extension.format
-import br.com.stone.emeraldcomponents.extension.month
-import kotlinx.android.synthetic.main.widget_date.view.*
+import kotlinx.android.synthetic.main.widget_date_filter.view.*
 import org.junit.Assert
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
-import org.robolectric.shadows.ShadowDialog
 import java.util.Calendar
 
 /**
@@ -27,79 +26,84 @@ import java.util.Calendar
 @RunWith(RobolectricTestRunner::class)
 class EmeraldDateFilterTest {
 
-    private lateinit var view: EmeraldDateSelector
+    private lateinit var view: EmeraldDateFilter
 
     private lateinit var activity: FragmentActivity
 
     @Before
     fun setup() {
         activity = Robolectric.buildActivity(FragmentActivity::class.java).create().get()
-        view = EmeraldDateSelector(activity)
+        view = EmeraldDateFilter(activity)
     }
 
     @Test
-    fun testInstanceWithContext() {
+    fun `Should create instance`() {
         Assert.assertNotNull(view)
     }
 
     @Test
-    fun testInstanceWithAttributeSet() {
+    fun `Should create instance with attribute set`() {
         val attrs = Robolectric.buildAttributeSet().build()
-        val view = EmeraldDateSelector(activity, attrs)
+        val view = EmeraldDateFilter(activity, attrs)
         Assert.assertNotNull(view)
     }
 
     @Test
-    fun testSetDate() {
-        val pattern = "MMMM, yyyy"
+    fun `Should set not personalized filter `() {
         val date = Calendar.getInstance()
-        view.date = date
-        assertEquals(date.format(pattern), view.emeraldTextDate.text)
+        view.setFilter(EmeraldDateFilterOptions.TODAY)
+        assertEquals(View.GONE, view.emeraldDateFilterExpandConstraint.visibility)
+        assertEquals(View.VISIBLE, view.emeraldTextDateRangeSubtitle.visibility)
+        assertEquals(EmeraldDateFilterOptions.TODAY, view.currentFilterType)
+        assertEquals(view.startDate, date)
+        assertEquals(view.endDate, date)
     }
 
     @Test
-    fun testSetDateChangedListener() {
-        val mockList = Mockito.mock(List::class.java)
-        view.setDateChangedListener { mockList.size }
-        view.emeraldImgRightArrowDate.performClick()
-        Mockito.verify(mockList).size
+    fun `Should set personalized filter`() {
+        val startDate = Calendar.getInstance()
+        view.setFilter(EmeraldDateFilterOptions.PERSONALIZED)
+        assertEquals(View.VISIBLE, view.emeraldDateFilterExpandConstraint.visibility)
+        assertEquals(View.GONE, view.emeraldTextDateRangeSubtitle.visibility)
+        assertEquals(EmeraldDateFilterOptions.PERSONALIZED, view.currentFilterType)
+        assertEquals("Personalizado", view.emeraldTextDateLabel.text)
+        assertEquals(startDate.format("dd/MM/yy"), view.emeraldDateFilterStartDate.text)
+        assertEquals(startDate.format("dd/MM/yy"), view.emeraldDateFilterEndDate.text)
     }
 
     @Test
-    fun testOnRightArrowClick() {
-        val month = 10
-        view.date.set(Calendar.DAY_OF_MONTH, 1)
-        view.date.set(Calendar.MONTH, month)
-        view.emeraldImgRightArrowDate.performClick()
-        assertEquals(month + 1, view.date.month())
+    fun `Should set custom filter`() {
+        val startDate = Calendar.getInstance()
+        val endDate = Calendar.getInstance().apply { add(Calendar.DAY_OF_YEAR, 1) }
+        view.setCustomFilter(startDate, endDate)
+        assertEquals(View.VISIBLE, view.emeraldDateFilterExpandConstraint.visibility)
+        assertEquals(View.GONE, view.emeraldTextDateRangeSubtitle.visibility)
+        assertEquals(EmeraldDateFilterOptions.PERSONALIZED, view.currentFilterType)
+        assertEquals("Personalizado", view.emeraldTextDateLabel.text)
+        assertEquals(startDate.format("dd/MM/yy"), view.emeraldDateFilterStartDate.text)
+        assertEquals(endDate.format("dd/MM/yy"), view.emeraldDateFilterEndDate.text)
     }
 
     @Test
-    fun testOnLeftArrowClick() {
-        val month = 10
-        view.date.set(Calendar.DAY_OF_MONTH, 1)
-        view.date.set(Calendar.MONTH, month)
-        view.emeraldImgLeftArrowDate.performClick()
-        assertEquals(month - 1, view.date.month())
+    fun `Should validate date range`() {
+        view.maxDaysRange = 2
+        val startDate = Calendar.getInstance()
+        val endDate = Calendar.getInstance().apply { add(Calendar.DAY_OF_YEAR, 1) }
+        assertTrue(view.isValidDateRange(startDate, endDate))
     }
 
     @Test
-    fun testOnDateTitleClick() {
-        view.emeraldTextDate.performClick()
-        Assert.assertTrue(ShadowDialog.getLatestDialog() is DatePickerDialog)
+    fun `Should not validate date range case endDate before startDate`() {
+        val startDate = Calendar.getInstance()
+        val endDate = Calendar.getInstance().apply { add(Calendar.DAY_OF_YEAR, -1) }
+        assertFalse(view.isValidDateRange(startDate, endDate))
     }
 
     @Test
-    fun testShowArrows() {
-        view.showArrows = true
-        assertEquals(View.VISIBLE, view.emeraldImgLeftArrowDate.visibility)
-        assertEquals(View.VISIBLE, view.emeraldImgRightArrowDate.visibility)
-    }
-
-    @Test
-    fun testHideArrows() {
-        view.showArrows = false
-        assertEquals(View.GONE, view.emeraldImgLeftArrowDate.visibility)
-        assertEquals(View.GONE, view.emeraldImgRightArrowDate.visibility)
+    fun `Should not validate date range case range bigger then maxRange`() {
+        view.maxDaysRange = 9
+        val startDate = Calendar.getInstance()
+        val endDate = Calendar.getInstance().apply { add(Calendar.DAY_OF_YEAR, 10) }
+        assertFalse(view.isValidDateRange(startDate, endDate))
     }
 }