+++ emeraldcomponents/src/test/java/br/com/stone/emeraldcomponents/basics/pager/tabs/EmeraldTabPagerTest.kt
@@ -1,6 +1,6 @@
 package br.com.stone.emeraldcomponents.basics.pager.tabs
 
-import android.support.v4.app.FragmentActivity
+import androidx.fragment.app.FragmentActivity
 import br.com.stone.emeraldcomponents.R
 import br.com.stone.emeraldcomponents.basic.pager.bullet.EmeraldBulletPager
 import br.com.stone.emeraldcomponents.basic.pager.tabs.EmeraldTabPager